NGINT-152: fixed artisan commands timeout

See merge request michelle-coates/psf-customer-portal/psi-customer-portal-backend!116

+++ config/artisan.php
@@ -1,5 +1,5 @@
 <?php
 
 return [
-    'timeout_seconds' => env('ARTISAN_TIMEOUT_SECONDS', 3600),
+    'timeout_seconds' => (int) env('ARTISAN_TIMEOUT_SECONDS', 3600),
 ];
